+++ tests/Unit/FeaturePolicyTest.php
@@ -4,7 +4,11 @@ use PHPUnit\Framework\TestCase;
 final class FeaturePolicyTest extends TestCase {
 	public function test_restricted_policy(): void {
 		$config = tragofone_feature_policy::configuration(
-			['extension' => '1001', 'password' => 'secret', 'domain_name' => 'pbx.test'],
+			[
+				'extension' => '1001', 'password' => 'secret', 'domain_name' => 'pbx.test',
+				'call_timeout' => '45', 'emergency_caller_id_number' => '+1 (415) 555-0911',
+				'voicemail_enabled' => 'true',
+			],
 			['sip_server' => '', 'sip_port' => 5061, 'sip_protocol' => 'tls', 'voicemail_code' => '*97'],
 			['+14155550100', '+14155550101']
 		);
@@ -13,6 +17,9 @@ final class FeaturePolicyTest extends TestCase {
 		self::assertSame('pbx.test', $config['Sip']['sip_auth_outboundProxyServer']);
 		self::assertSame('5061', $config['Sip']['sip_auth_outboundProxyPort']);
 		self::assertSame('*97', $config['Call']['call_onetouch_voicemailNumber']);
+		self::assertSame('45', $config['Call']['call_noAnswerTimeout']);
+		self::assertSame('+14155550911', $config['emergency']['emergency_numbers']);
+		self::assertSame('TRUE', $config['voicemail']['voicemail_status']);
 		self::assertSame('FALSE', $config['IM']['im_status']);
 		self::assertSame('FALSE', $config['Sms']['sms_status']);
 		self::assertSame('FALSE', $config['Video']['video_enableVideo']);
@@ -20,6 +27,17 @@ final class FeaturePolicyTest extends TestCase {
 		self::assertSame('FALSE', $config['configurations']['configurations_dndVisibility']);
 		self::assertSame('FALSE', $config['blf']['autoenable_blf']);
 		self::assertSame('FALSE', $config['zoom']['zoom_status']);
+	}
+
+	public function test_clears_emergency_number_and_disables_voicemail_from_fusionpbx(): void {
+		$config = tragofone_feature_policy::configuration(
+			['extension'=>'AB12','password'=>'secret','domain_name'=>'pbx.test','call_timeout'=>null,'emergency_caller_id_number'=>'','voicemail_enabled'=>'false'],
+			['sip_server'=>'','sip_port'=>5061,'sip_protocol'=>'tls','voicemail_code'=>'*97'], []
+		);
+		self::assertSame('AB12', $config['Sip']['sip_extension']);
+		self::assertSame('30', $config['Call']['call_noAnswerTimeout']);
+		self::assertSame('', $config['emergency']['emergency_numbers']);
+		self::assertSame('FALSE', $config['voicemail']['voicemail_status']);
 	}
 
 	public function test_uses_explicit_outbound_proxy_when_configured(): void {